fix: systemd socket support build handling for windows
